Refactor nodes
