+++ enclave-manager/web/packages/app/src/emui/enclaves/components/enclaveBuilder/nodes/KurtosisServiceNode.tsx
@@ -2,21 +2,21 @@ import { Flex, Tab, TabList, TabPanel, TabPanels, Tabs } from "@chakra-ui/react"
 import { isDefined } from "kurtosis-ui-components";
 import { memo } from "react";
 import { NodeProps } from "reactflow";
-import { BooleanArgumentInput } from "../form/BooleanArgumentInput";
-import { DictArgumentInput } from "../form/DictArgumentInput";
-import { IntegerArgumentInput } from "../form/IntegerArgumentInput";
-import { KurtosisFormControl } from "../form/KurtosisFormControl";
-import { ListArgumentInput } from "../form/ListArgumentInput";
-import { StringArgumentInput } from "../form/StringArgumentInput";
-import { KurtosisFormInputProps } from "../form/types";
-import { ImageConfigInput } from "./input/ImageConfigInput";
-import { MentionStringArgumentInput } from "./input/MentionStringArgumentInput";
-import { MountArtifactFileInput } from "./input/MountArtifactFileInput";
-import { PortConfigurationField } from "./input/PortConfigurationInput";
-import { validateName } from "./input/validators";
+import { BooleanArgumentInput } from "../../form/BooleanArgumentInput";
+import { DictArgumentInput } from "../../form/DictArgumentInput";
+import { IntegerArgumentInput } from "../../form/IntegerArgumentInput";
+import { KurtosisFormControl } from "../../form/KurtosisFormControl";
+import { ListArgumentInput } from "../../form/ListArgumentInput";
+import { StringArgumentInput } from "../../form/StringArgumentInput";
+import { KurtosisFormInputProps } from "../../form/types";
+import { ImageConfigInput } from "../input/ImageConfigInput";
+import { MentionStringArgumentInput } from "../input/MentionStringArgumentInput";
+import { MountArtifactFileInput } from "../input/MountArtifactFileInput";
+import { PortConfigurationField } from "../input/PortConfigurationInput";
+import { validateName } from "../input/validators";
+import { KurtosisFileMount, KurtosisPort, KurtosisServiceNodeData } from "../types";
+import { useVariableContext } from "../VariableContextProvider";
 import { KurtosisNode } from "./KurtosisNode";
-import { KurtosisFileMount, KurtosisPort, KurtosisServiceNodeData } from "./types";
-import { useVariableContext } from "./VariableContextProvider";
 
 export const KurtosisServiceNode = memo(
   ({ id, selected }: NodeProps) => {
